feat: 공유 기록 오픈그래프 적용
